+++ discord/objects/components/__init__.py
@@ -1,3 +1,4 @@
 from .checkbox import CheckboxComponent
 from .checkbox_group import CheckboxGroupComponent
 from .checkbox_group_option import CheckboxGroupOption
+from .radio_group import RadioGroupComponent